Implement default file path pattern for tables component
