+++ hydromt/components/tables.py
@@ -14,12 +14,13 @@ import pandas as pd
 
 from hydromt.components.base import ModelComponent
 from hydromt.hydromt_step import hydromt_step
-from hydromt.utils.constants import DEFAULT_TABLE_FILENAME
 
 if TYPE_CHECKING:
     from hydromt.models.model import Model
 
 __all__ = ["TablesComponent"]
+
+_DEFAULT_TABLE_FILENAME = "tables/{name}.csv"
 
 
 class TablesComponent(ModelComponent):
@@ -28,10 +29,7 @@ class TablesComponent(ModelComponent):
     It is well suited to represent non-geospatial tabular model data.
     """
 
-    def __init__(
-        self,
-        model: "Model",
-    ):
+    def __init__(self, model: "Model", default_filename: str = _DEFAULT_TABLE_FILENAME):
         """Initialize a TablesComponent.
 
         Parameters
@@ -40,6 +38,7 @@ class TablesComponent(ModelComponent):
             HydroMT model instance
         """
         self._data: Optional[Dict[str, Union[pd.DataFrame, pd.Series]]] = None
+        self._filename = default_filename
         super().__init__(model=model)
 
     @property
@@ -59,9 +58,10 @@ class TablesComponent(ModelComponent):
                 self.read()
 
     @hydromt_step
-    def write(self, fn: str = DEFAULT_TABLE_FILENAME, **kwargs) -> None:
+    def write(self, filename: Optional[str] = None, **kwargs) -> None:
         """Write tables at <root>/tables."""
         self._root._assert_write_mode()
+        fn = filename or self._filename
         if len(self.data) > 0:
             self._model.logger.info("Writing table files.")
             local_kwargs = {"index": False, "header": True, "sep": ","}
@@ -74,11 +74,12 @@ class TablesComponent(ModelComponent):
             self._model.logger.debug("No tables found, skip writing.")
 
     @hydromt_step
-    def read(self, fn: str = DEFAULT_TABLE_FILENAME, **kwargs) -> None:
+    def read(self, filename: Optional[str] = None, **kwargs) -> None:
         """Read table files at <root>/tables and parse to dict of dataframes."""
         self._root._assert_read_mode()
         self._initialize_tables(skip_read=True)
         self._model.logger.info("Reading model table files.")
+        fn = filename or self._filename
         filenames = glob.glob(join(self._root.path, fn.format(name="*")))
         if len(filenames) > 0:
             for fn in filenames:
