+++ hydromt/utils/__init__.py
@@ -1,6 +1,5 @@
 """Public helper functions that (pulgin) developers can use."""
 
-from hydromt.utils.constants import DEFAULT_TABLE_FILENAME
 from hydromt.utils.deep_merge import deep_merge
 
-__all__ = ["deep_merge", "DEFAULT_TABLE_FILENAME"]
+__all__ = ["deep_merge"]
