--- hydromt/utils/constants.py
@@ -1,3 +1,0 @@
-"""Publically available constants for hydromt."""
-
-DEFAULT_TABLE_FILENAME = "tables/{name}.csv"
